# RFC: Reduce memory allocations [Do not merge]