--- a/benchmark.ps1
+++ b/benchmark.ps1
@@ -1,0 +1,11 @@
+param(
+	[Parameter(Mandatory=$True)]$Benchmark = 'CommandBenchmark'
+)
+
+$ErrorActionPreference = 'Stop'
+
+
+
+dotnet build .\src\Perf\Perf.csproj --configuration 'Release'
+& .\src\Perf\bin\Release\net7.0\Perf.exe --filter "*$($Benchmark)*"
+