--- a/src/FirebirdSql.Data.FirebirdClient/Client/Managed/Version16/GdsBatch.cs
+++ b/src/FirebirdSql.Data.FirebirdClient/Client/Managed/Version16/GdsBatch.cs
@@ -44,7 +44,7 @@ internal class GdsBatch : BatchBase
 		Database.Xdr.Write(IscCodes.op_batch_create);
 		Database.Xdr.Write(_statement.Handle); // p_batch_statement
 		var blr = _statement.Parameters.ToBlr();
-		Database.Xdr.WriteBuffer(blr.Data); // p_batch_blr
+		Database.Xdr.WriteBuffer(blr); // p_batch_blr
 		Database.Xdr.Write(blr.Length); // p_batch_msglen
 		var pb = _statement.CreateBatchParameterBuffer();
 		if (_statement.ReturnRecordsAffected)
@@ -96,8 +96,8 @@ internal class GdsBatch : BatchBase
 
 		await Database.Xdr.WriteAsync(IscCodes.op_batch_create, cancellationToken).ConfigureAwait(false);
 		await Database.Xdr.WriteAsync(_statement.Handle, cancellationToken).ConfigureAwait(false); // p_batch_statement
-		var blr = _statement.Parameters.ToBlr();
-		await Database.Xdr.WriteBufferAsync(blr.Data, cancellationToken).ConfigureAwait(false); // p_batch_blr
+		var blr = _statement.Parameters.ToBlrMemory();
+		await Database.Xdr.WriteBufferAsync(blr, cancellationToken).ConfigureAwait(false); // p_batch_blr
 		await Database.Xdr.WriteAsync(blr.Length, cancellationToken).ConfigureAwait(false); // p_batch_msglen
 		var pb = _statement.CreateBatchParameterBuffer();
 		if (_statement.ReturnRecordsAffected)