--- a/src/FirebirdSql.Data.FirebirdClient/Client/Managed/Version16/GdsBatch.cs
+++ b/src/FirebirdSql.Data.FirebirdClient/Client/Managed/Version16/GdsBatch.cs
@@ -43,9 +43,9 @@ internal class GdsBatch : BatchBase
 
 		Database.Xdr.Write(IscCodes.op_batch_create);
 		Database.Xdr.Write(_statement.Handle); // p_batch_statement
-		var blr = _statement.Parameters.ToBlr();
+		var blr = _statement.Parameters.ToBlrSpan(out var blrMsgLen);
 		Database.Xdr.WriteBuffer(blr); // p_batch_blr
-		Database.Xdr.Write(blr.Length); // p_batch_msglen
+		Database.Xdr.Write(blrMsgLen); // p_batch_msglen
 		var pb = _statement.CreateBatchParameterBuffer();
 		if (_statement.ReturnRecordsAffected)
 		{
@@ -96,9 +96,9 @@ internal class GdsBatch : BatchBase
 
 		await Database.Xdr.WriteAsync(IscCodes.op_batch_create, cancellationToken).ConfigureAwait(false);
 		await Database.Xdr.WriteAsync(_statement.Handle, cancellationToken).ConfigureAwait(false); // p_batch_statement
-		var blr = _statement.Parameters.ToBlrMemory();
+		var blr = _statement.Parameters.ToBlrMemory(out var blrMsgLen);
 		await Database.Xdr.WriteBufferAsync(blr, cancellationToken).ConfigureAwait(false); // p_batch_blr
-		await Database.Xdr.WriteAsync(blr.Length, cancellationToken).ConfigureAwait(false); // p_batch_msglen
+		await Database.Xdr.WriteAsync(blrMsgLen, cancellationToken).ConfigureAwait(false); // p_batch_msglen
 		var pb = _statement.CreateBatchParameterBuffer();
 		if (_statement.ReturnRecordsAffected)
 		{