--- a/src/FirebirdSql.Data.FirebirdClient/Client/Native/Marshalers/XSQLDA.cs
+++ b/src/FirebirdSql.Data.FirebirdClient/Client/Native/Marshalers/XSQLDA.cs
@@ -15,20 +15,7 @@
 
 //$Authors = Carlos Guzman Alvarez, Jiri Cincura (jiri@cincura.net)
 
-using System.Runtime.InteropServices;
-
 namespace FirebirdSql.Data.Client.Native.Marshalers;
-
-[StructLayout(LayoutKind.Sequential)]
-internal struct XSQLDA
-{
-	public short version;
-	[MarshalAs(UnmanagedType.ByValTStr, SizeConst = 8)]
-	public string sqldaid;
-	public int sqldabc;
-	public short sqln;
-	public short sqld;
-}
 
 internal unsafe struct XSQLDA_STRUCT
 {