… src/FirebirdSql.Data.FirebirdClient/Common/DbField.cs
… src/FirebirdSql.Data.FirebirdClient/Common/Descriptor.cs
… src/FirebirdSql.Data.FirebirdClient/Common/TypeDecoder.cs
… src/FirebirdSql.Data.FirebirdClient/FirebirdSql.Data.FirebirdClient.csproj
… src/Perf/Perf.csproj
… src/Perf/PerformanceBenchmark.cs
… src/Perf/Program.cs