@fdcastel 2023-10-01
Done! Please review.

P.S.: Remember that for these changes I had to remove net48, netstandard2.0 and net5.0 support.