@fdcastel 2023-10-02
Sorry, @cincuranet -- Something in [this commit](https://github.com/cincuranet/FirebirdSql.Data.FirebirdClient/pull/122/commits/a497ca8d11613d12a20e7ea25e2eb44f9427f04b) is broken but I'm not able to debug it anymore for now. 

I will review this in the next few days.

Meanwhile, the commits up to [this](https://github.com/cincuranet/FirebirdSql.Data.FirebirdClient/pull/122/commits/7a3f47410cf4d82626dc364444e126fe5dd06aa0) are fine.