@cincuranet 2024-08-26
Overall this looks fine. Just create separate PRs, with minimal changes in related pieces - i.e. changes in `XSQLDA` and around and nothing else. That way it will be easier to review and merge.

I'll close this now and we can discuss specifics in respective PRs.